# Optimise `HashingEncoder` for both large and small dataframes