@PaulWestenthanner 2023-10-10
one more question: you've removed all the data-lock stuff for multi-processing, right. Was this not necessary?