@bkhant1 2023-10-20
Hi @PaulWestenthanner ! Thanks for your review!

> one more question: you've removed all the data-lock stuff for multi-processing, right. Was this not necessary?

It was necessary because all processes were writting to the same queue. With shared memory they write to their own section of the shared memory, so no concurrency to manage there.

I think I replied to all your comments/addressed them. 

The one outstanding thing is that you suggested:
> Wouldn't it be easier if the hash_chunk function would hash a chunk and return an array. Then it wouldn't need the shm_result and shm_offset parameters (what does shm stand for btw?). Then you'd just concatenate all the chunks in the end?

It's a little more complex than that because `return` doesn't really work accross processes (hence the use of queues or shared memory), but I remembered that python has a nice API, [ProcessPoolExecutor](https://docs.python.org/3/library/concurrent.futures.html#processpoolexecutor), (that uses queues behind the scenes) to make it look like it's possible to return something from a process. 

Here's [a PR](https://github.com/bkhant1/category_encoders/compare/all_optis...bkhant1:category_encoders:multiproc_pool?expand=1) in my fork showing what I tried. It could be a little slower but the code is cleaner. I think I should add this commit to this PR but let me know what you think!

|Baseline                                                |Before review                                    |After review                                     |Multiproc pool instead of SHM                    |FIELD5                                           |
|--------------------------------------------------------|-------------------------------------------------|-------------------------------------------------|-------------------------------------------------|-------------------------------------------------|
|n_rows=30 n_features=3 n_components=10 n_process=4      |3.12 s ± 20.2 ms per loop (mean ± std. dev. of...|53.9 ms ± 443 µs per loop (mean ± std. dev. of...|55 ms ± 1.21 ms per loop (mean ± std. dev. of ...|59.8 ms ± 989 µs per loop (mean ± std. dev. of...|
|n_rows=30 n_features=3 n_components=10 n_process=1      |1.1 s ± 17 ms per loop (mean ± std. dev. of 7 ...|1.56 ms ± 45.5 µs per loop (mean ± std. dev. o...|1.5 ms ± 44.8 µs per loop (mean ± std. dev. of...|1.52 ms ± 39.8 µs per loop (mean ± std. dev. o...|
|n_rows=30 n_features=3 n_components=100 n_process=1     |1.1 s ± 14.5 ms per loop (mean ± std. dev. of ...|1.52 ms ± 25.2 µs per loop (mean ± std. dev. o...|1.77 ms ± 317 µs per loop (mean ± std. dev. of...|1.49 ms ± 21.7 µs per loop (mean ± std. dev. o...|
|n_rows=10000 n_features=10 n_components=10 n_process=4  |4.75 s ± 36 ms per loop (mean ± std. dev. of 7...|99.3 ms ± 649 µs per loop (mean ± std. dev. of...|122 ms ± 13 ms per loop (mean ± std. dev. of 7...|113 ms ± 992 µs per loop (mean ± std. dev. of ...|
|n_rows=10000 n_features=10 n_components=10 n_process=1  |1.51 s ± 13.5 ms per loop (mean ± std. dev. of...|156 ms ± 2.67 ms per loop (mean ± std. dev. of...|165 ms ± 17.6 ms per loop (mean ± std. dev. of...|146 ms ± 1.79 ms per loop (mean ± std. dev. of...|
|n_rows=100000 n_features=10 n_components=10 n_process=4 |5.53 s ± 30.6 ms per loop (mean ± std. dev. of...|584 ms ± 6.74 ms per loop (mean ± std. dev. of...|576 ms ± 25.2 ms per loop (mean ± std. dev. of...|572 ms ± 12.4 ms per loop (mean ± std. dev. of...|
|n_rows=100000 n_features=10 n_components=10 n_process=1 |5.22 s ± 55.3 ms per loop (mean ± std. dev. of...|1.55 s ± 11 ms per loop (mean ± std. dev. of 7...|1.48 s ± 12.6 ms per loop (mean ± std. dev. of...|1.46 s ± 32.7 ms per loop (mean ± std. dev. of...|
|n_rows=1000000 n_features=50 n_components=10 n_process=4|53.2 s ± 2.35 s per loop (mean ± std. dev. of ...|23.5 s ± 256 ms per loop (mean ± std. dev. of ...|22.3 s ± 200 ms per loop (mean ± std. dev. of ...|25.6 s ± 1.6 s per loop (mean ± std. dev. of 7...|