@PaulWestenthanner 2023-10-28
Thanks for the changes. Looks really cool now. 
I'd prefer the clean `ProcessPoolExecutor` implementation. Then we'll be able to merge!