@PaulWestenthanner 2023-11-11
Looks good! Thanks for this improvement! Makes it both more readable and faster